+++ node/examples/streaming.ts
@@ -31,9 +31,9 @@ const stream = client.recognize({
     sampleRateHertz: getWaveSampleRate(fileName),
     partialResults: true,
     model: process.env.MODEL || 'generic-small-en-us-0.15',
-    // The locale is optional and only relevant for multilingual models which autodetect the language if not specified explicitly.
+    // The locale is optional and only relevant for multilingual multitask models which autodetect the language if not specified explicitly.
     locale: process.env.LOCALE || '', // e.g. 'en
-    // Optional - Can be used for multilingual models to add a glossary with task-specific vocabulary (e.g. product names) that would otherwise be unlikely to be recognized. Can also be used to steer the recognition towards a specific writing style.
+    // Optional - Can be used for multitask models to add a glossary with task-specific vocabulary (e.g. product names) that would otherwise be unlikely to be recognized. Can also be used to steer the recognition towards a specific writing style.
     prompt: process.env.MODEL_PROMPT,
   }
 })